cookiecutter-cli-app/{{cookiecutter.directory_name}}/requirements_dev.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
text-unidecode==1.3
urllib3==1.26.14
